+++ CHANGELOG.rst
@@ -16,6 +16,7 @@
     * Support Excel file format for `inventory`, `gen` and `attrib`
     * Add 'spdx_license_key' support
     * Add option to save error log in `check` command
+    * New `gen_license` option
 
 2021-04-02
     Release 6.0.0